Refactor bcheck package to data package.
