+++ src/main/java/data/BCheck.java
@@ -1,4 +1,4 @@
-package bcheck;
+package data;
 
 import java.io.IOException;
 import java.nio.file.Path;
