+++ src/main/java/data/BCheckFactory.java
@@ -1,4 +1,4 @@
-package bcheck;
+package data;
 
 import logging.Logger;
