+++ src/main/java/data/BCheckFilter.java
@@ -1,4 +1,4 @@
-package bcheck;
+package data;
 
 public class BCheckFilter implements ItemFilter<BCheck> {
     @Override
