+++ src/main/java/data/ItemFilter.java
@@ -1,4 +1,4 @@
-package bcheck;
+package data;
 
 public interface ItemFilter<T extends Item> {
     boolean filter(T item, String searchTerm);
