+++ src/main/java/data/Tags.java
@@ -1,4 +1,4 @@
-package bcheck;
+package data;
 
 import java.util.List;
 import java.util.Objects;
