+++ src/main/java/repository/GitHubRepository.java
@@ -1,8 +1,8 @@
 package repository;
 
-import bcheck.BCheck;
-import bcheck.BCheckFactory;
 import client.GitHubClient;
+import data.BCheck;
+import data.BCheckFactory;
 import file.finder.BCheckFileFinder;
 import file.temp.TempFileCreator;
 import file.zip.ZipExtractor;
